Combine unit tests for calendar utils
